requirements.txt
PyQt6==6.5.0
Pillow==10.1.0
